TST: fix 32-bit/python < 2.7 test failures

+++ pandas/src/datetime.pyx
@@ -213,7 +213,8 @@ cdef class _Timestamp(datetime):
             return datetime.__sub__(self, other)
 
     def _get_field(self, field):
-        out = fast_field_accessor(np.array([self.value]), field)
+        out = fast_field_accessor(np.array([self.value], dtype=np.int64),
+                                  field)
         return out[0]
 
 # lightweight C object to hold datetime & int64 pair
